requirements.txt
Django==2.2.5
django-crispy-forms==1.7.2
djangorestframework==3.11.0
djangorestframework-simplejwt==1.2
ecdsa==0.15
Pillow==7.0.0
pyasn1==0.4.8
python-jose==3.1.0
pytz==2019.2
rsa==4.0
six==1.14.0
sqlparse==0.3.0
